# kubegram-server/docs/swagger.yaml
basePath: /api/public/v1
definitions:
...
  github_com_kubegram_kubegram-server_src_models.Organization:
    properties:
      companyID:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      teams:
        items:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Team'
        type: array
      updatedAt:
        type: string
    type: object
  github_com_kubegram_kubegram-server_src_models.Team:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      organizationID:
        type: integer
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.User'
        type: array
    type: object
  github_com_kubegram_kubegram-server_src_models.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      provider:
        description: OAuth linkage
        type: string
      provider_id:
        type: string
      role:
        description: '"admin" | "manager" | "team_member"'
        type: string
      teamID:
        type: integer
      updated_at:
        type: string
      updatedAt:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  github_com_kubegram_kubegram-server_src_models.OrganizationManifest:
    properties:
      apiVersion:
        type: string
      kind:
        type: string
      metadata:
        type: object
        properties:
          name:
            type: string
      spec:
        type: object
        properties:
          companyID:
            type: integer
    type: object
  github_com_kubegram_kubegram-server_src_models.TeamManifest:
    properties:
      apiVersion:
        type: string
      kind:
        type: string
      metadata:
        type: object
        properties:
          name:
            type: string
      spec:
        type: object
        properties:
          organizationID:
            type: integer
    type: object
...
host: localhost:8090
securityDefinitions:
  OAuth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: /auth/{provider}/login
    tokenUrl: /auth/{provider}/callback
    scopes:
      admin: Admin access for internal operations
      read: Read access
      write: Write access
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Kubegram application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kubegram API
  version: "1.0"
paths:
...
  /api/v1/organizations:
    get:
      description: Get a list of all organizations or look up a specific organization by team ID
      parameters:
      - description: Team ID to look up associated organization
        in: query
        name: teamId
        required: false
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            oneOf:
            - $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Organization'
            - items:
                $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Organization'
              type: array
        "400":
          description: Bad Request - Invalid teamId parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found - Team not found or team has no organization
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all organizations or look up organization by team ID
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Create a new organization
      parameters:
      - description: Organization object
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Organization'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Organization'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a organization
      tags:
      - organizations
  /api/v1/organizations/{id}:
    delete:
      description: Delete a organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a organization
      tags:
      - organizations
    get:
      description: Get a organization by its ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Organization'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a organization
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Update a organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      - description: Organization object
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Organization'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Organization'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a organization
      tags:
      - organizations
...
  /api/v1/teams:
    get:
      description: Get a list of all teams or look up a specific team by user ID
      parameters:
      - description: User ID to look up associated team
        in: query
        name: userId
        required: false
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            oneOf:
            - $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Team'
            - items:
                $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Team'
              type: array
        "400":
          description: Bad Request - Invalid userId parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found - User not found or user has no team
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all teams or look up team by user ID
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Create a new team
      parameters:
      - description: Team object
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Team'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a team
      tags:
      - teams
  /api/v1/teams/{id}:
    delete:
      description: Delete a team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a team
      tags:
      - teams
    get:
      description: Get a team by its ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a team
      tags:
      - teams
    put:
      consumes:
      - application/json
      description: Update a team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: Team object
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Team'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a team
      tags:
      - teams
...
  /api/v1/iac/organizations:
    post:
      description: Create a new organization resource using K8s manifest
      parameters:
      - description: Organization Manifest
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.OrganizationManifest'
      responses:
        "201":
          description: Created
      summary: Create Organization (IaC)
      tags:
      - organizations
  /api/v1/iac/organizations/{id}:
    put:
      description: Update a organization resource using K8s manifest
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      - description: Organization Manifest
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.OrganizationManifest'
      responses:
        "200":
          description: OK
      summary: Update Organization (IaC)
      tags:
      - organizations
    delete:
      description: Delete a organization resource (IaC)
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete Organization (IaC)
      tags:
      - organizations
  /api/v1/iac/teams:
    post:
      description: Create a new team resource using K8s manifest
      parameters:
      - description: Team Manifest
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.TeamManifest'
      responses:
        "201":
          description: Created
      summary: Create Team (IaC)
      tags:
      - teams
  /api/v1/iac/teams/{id}:
    put:
      description: Update a team resource using K8s manifest
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: Team Manifest
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/github_com_kubegram_kubegram-server_src_models.TeamManifest'
      responses:
        "200":
          description: OK
      summary: Update Team (IaC)
      tags:
      - teams
    delete:
      description: Delete a team resource (IaC)
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete Team (IaC)
      tags:
      - teams
...
swagger: "2.0"